Move client handler to a separate module
